HLK-LD2451/requirements.txt
pyserial
matplotlib
colorama
